# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Equipment:
    properties:
      count:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      img_url:
        type: string
      name:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  models.Member:
    properties:
      address:
        type: string
      birthday:
        type: string
      contact:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      instagram:
        type: string
      name:
        type: string
      nik:
        type: integer
      package:
        type: string
      trainer:
        $ref: '#/definitions/models.Trainer'
      trainer_id:
        type: integer
      updatedAt:
        type: string
      weight:
        type: integer
    type: object
  models.MemberBody:
    properties:
      address:
        type: string
      birthday:
        type: string
      contact:
        type: string
      email:
        type: string
      gender:
        type: string
      instagram:
        type: string
      name:
        type: string
      nik:
        type: integer
      package:
        type: string
      trainer_id:
        type: integer
      weight:
        type: integer
    type: object
  models.Trainer:
    properties:
      achievement:
        type: string
      address:
        type: string
      birthday:
        type: string
      contact:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      email:
        type: string
      experience:
        type: string
      gender:
        type: string
      id:
        type: integer
      img_url:
        type: string
      instagram:
        type: string
      members:
        items:
          $ref: '#/definitions/models.Member'
        type: array
      name:
        type: string
      nik:
        type: integer
      specialization:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: akram.firman@gmail.com
    name: API Support
  description: This is a swagger for Jagona Gym API
  title: Jagona Gym API
  version: 0.0.1
paths:
  /api/equipment:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Equipment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all equipments
      tags:
      - Equipment
    post:
      consumes:
      - multipart/form-data
      description: Creating new Equipment data
      parameters:
      - description: Nama Alat
        in: formData
        name: name
        required: true
        type: string
      - description: Jumlah Alat
        in: formData
        name: count
        required: true
        type: integer
      - description: 'Status Alat: ready atau broken'
        in: formData
        name: status
        required: true
        type: string
      - description: Foto Alat
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success create equipment
          schema:
            $ref: '#/definitions/models.Equipment'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Equipment
      tags:
      - Equipment
  /api/equipment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Equipment data by id
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success Delete Equipment Data
          schema:
            type: string
        "404":
          description: Data not Found!
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete Equipment
      tags:
      - Equipment
    get:
      consumes:
      - application/json
      description: Get a single Equipment data by ID
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Equipment'
        "404":
          description: Equipment not found
          schema:
            type: string
        "500":
          description: Failed to get equipment
          schema:
            type: string
      summary: Get a single Equipment
      tags:
      - Equipment
    put:
      consumes:
      - multipart/form-data
      description: Update Equipment data by ID
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nama Alat
        in: formData
        name: name
        required: true
        type: string
      - description: Jumlah Alat
        in: formData
        name: count
        required: true
        type: integer
      - description: 'Status Alat: ready atau broken'
        in: formData
        name: status
        required: true
        type: string
      - description: Foto Alat
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success create equipment
          schema:
            $ref: '#/definitions/models.Equipment'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update Equipment
      tags:
      - Equipment
  /api/member:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Member'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all members
      tags:
      - Member
    post:
      consumes:
      - application/json
      description: Creating new Member data
      parameters:
      - description: Data member
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.MemberBody'
      produces:
      - application/json
      responses:
        "200":
          description: Success create member
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Member
      tags:
      - Member
  /api/member/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Member data by id
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success Delete Member Data
          schema:
            type: string
        "404":
          description: Data not Found!
          schema:
            type: string
      summary: Delete Member
      tags:
      - Member
    get:
      consumes:
      - application/json
      description: Get a single Member data by ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "404":
          description: Member not found
          schema:
            type: string
        "500":
          description: Failed to get trainer
          schema:
            type: string
      summary: Get a single Member
      tags:
      - Member
    put:
      consumes:
      - application/json
      description: Update a specific Member data
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Member data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.MemberBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Member not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update Member
      tags:
      - Member
  /api/trainer:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Trainer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all trainers
      tags:
      - Trainer
    post:
      consumes:
      - multipart/form-data
      description: Creating new Trainer data
      parameters:
      - description: Name trainer
        in: formData
        name: name
        required: true
        type: string
      - description: NIK trainer
        in: formData
        name: nik
        required: true
        type: integer
      - description: Tanggal Lahir trainer
        in: formData
        name: birthday
        required: true
        type: string
      - description: Email trainer
        in: formData
        name: email
        required: true
        type: string
      - description: Kontak trainer
        in: formData
        name: contact
        required: true
        type: string
      - description: Instagram trainer
        in: formData
        name: instagram
        type: string
      - description: Alamat trainer
        in: formData
        name: address
        required: true
        type: string
      - description: Gender trainer
        in: formData
        name: gender
        required: true
        type: string
      - description: Deskripsi singkat trainer
        in: formData
        name: description
        type: string
      - description: Pengalaman trainer
        in: formData
        name: experience
        type: string
      - description: Spesialisasi trainer
        in: formData
        name: specialization
        required: true
        type: string
      - description: Pencapaian/Sertifikasi/Lisensi trainer
        in: formData
        name: achievement
        type: string
      - description: Image trainer
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success create trainer
          schema:
            $ref: '#/definitions/models.Trainer'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Trainer
      tags:
      - Trainer
  /api/trainer/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Trainer data by id
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success Delete Trainer Data
          schema:
            type: string
        "404":
          description: Data not Found!
          schema:
            type: string
      summary: Delete Trainer
      tags:
      - Trainer
    get:
      consumes:
      - application/json
      description: Get a single Trainer data by ID
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trainer'
        "404":
          description: Trainer not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a single Trainer
      tags:
      - Trainer
    put:
      consumes:
      - multipart/form-data
      description: Update a specific Trainer data
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name trainer
        in: formData
        name: name
        required: true
        type: string
      - description: NIK trainer
        in: formData
        name: nik
        required: true
        type: integer
      - description: Tanggal Lahir trainer
        in: formData
        name: birthday
        required: true
        type: string
      - description: Email trainer
        in: formData
        name: email
        required: true
        type: string
      - description: Kontak trainer
        in: formData
        name: contact
        required: true
        type: string
      - description: Instagram trainer
        in: formData
        name: instagram
        required: true
        type: string
      - description: Alamat trainer
        in: formData
        name: address
        required: true
        type: string
      - description: Gender trainer
        in: formData
        name: gender
        required: true
        type: string
      - description: Deskripsi singkat trainer
        in: formData
        name: description
        required: true
        type: string
      - description: Pengalaman trainer
        in: formData
        name: experience
        required: true
        type: string
      - description: Spesialisasi trainer
        in: formData
        name: specialization
        required: true
        type: string
      - description: Pencapaian/Sertifikasi/Lisensi trainer
        in: formData
        name: achievement
        required: true
        type: string
      - description: Image trainer
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trainer'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Trainer not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update Trainer
      tags:
      - Trainer
swagger: "2.0"